+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/crediya/r2dbc/entity/UserEntity.java
@@ -6,6 +6,8 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 
+import java.math.BigDecimal;
+
 @Table(name = "users")
 @AllArgsConstructor
 @NoArgsConstructor
@@ -33,7 +35,7 @@ public class UserEntity {
     @Column("identification_number")
     private String identificationNumber;
     @Column("base_salary")
-    private Double baseSalary;
+    private BigDecimal baseSalary;
     @Column("rol_id")
     private String rolId;
 }
